Fix another harmless compiler warning.  Get the MSVC makefile working with unionvtab.
FossilOrigin-Name: d49639c208ac366f6c675ac5671accc2a597558b62a51520194276c61f32d406

+++ ext/misc/unionvtab.c
@@ -133,7 +133,7 @@ static void *unionMalloc(int *pRc, int nByte){
 static char *unionStrdup(int *pRc, const char *zIn){
   char *zRet = 0;
   if( zIn ){
-    int nByte = strlen(zIn) + 1;
+    int nByte = (int)strlen(zIn) + 1;
     zRet = unionMalloc(pRc, nByte);
     if( zRet ){
       memcpy(zRet, zIn, nByte);
